v 1.0 of the plugin
The current iteration of the WordPress plugin creates a custom post type "Messages" and with a shortcode, you can create a frontend form to create "message" posts.
Additionally, an e-mail is being sent to all administrators containing everything that was submitted via the form. So basically, the whole plugin allows you to use Advanced Custom Fields (ACF) as a contact form without the need for additional plugins.